microservices/audio-generation-service/requirements.txt
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
google-cloud-storage==2.14.0
google-cloud-texttospeech==2.14.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.23.3
python-json-logger==2.0.7
pyttsx3==2.90  # Fallback TTS engine
soundfile==0.12.1
numpy==1.24.3 